refactor(tcp)!: Refactor TCP connection limit to use queue backpressure
- Replaces manual connection counting and dropping with a queue-based backpressure model.
- MaxConnections is changed from a Int64 to an Int32 and sets the max size of the internal processing queue

Uses a new backpressure model by waiting for the queue to free space before returning back to the OS accept(), which applies backpressre back to the operating system backlog when accept threads are all waiting for space in the work queue.

+++ lib/Net.Transport.Tcp/src/TcpConfig.cs
@@ -68,9 +68,10 @@ namespace VNLib.Net.Transport.Tcp
         public required readonly int MaxRecvBufferData { get; init; }
 
         /// <summary>
-        /// The maximum number of allowed socket connections to this server
+        /// The maximum number of allowed socket connections to this server, 
+        /// before backpressure is applied down to the OS.
         /// </summary>
-        public required readonly long MaxConnections { get; init; }
+        public required readonly int MaxConnections { get; init; }
 
         /// <summary>
         /// The listener socket backlog count
